Made local visual components to display a dropdown
There are a few issues with  the current implementation
One critical issue with it is it's lack of portability. It requires
patches to be made to the clientside runtime directories in
order to interpret the type of the property editor and
properly append it.

A few hack-y solutions exist to fix it, however for now I will work
with this implementation and work on packaging it later.