Threaded metadata fetching should work now
In order to accomplish this, I had to create
a new class called Metadata. This class is needed
to unify data in order to override AsyncTask
cleanly.

(#1)
This results in a large amount of responsiveness
when using incremental fetching out of a list.
However the result occaisionally has some interesting
behavior. This commit needs some testing and documentation
of said behavior before release.